+++ dcmrt/libsrc/drtafs.cc
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2016, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Source file for class DRTAssigningFacilitySequence
  *
  *  Generated automatically from DICOM PS 3.3-2016e
- *  File created on 2016-11-23 14:23:36
+ *  File created on 2017-01-25 17:55:32
  *
  */
 
@@ -280,7 +280,7 @@ OFBool DRTAssigningFacilitySequence::isValid() const
 }
 
 
-unsigned long DRTAssigningFacilitySequence::getNumberOfItems() const
+size_t DRTAssigningFacilitySequence::getNumberOfItems() const
 {
     return SequenceOfItems.size();
 }
@@ -310,12 +310,12 @@ OFCondition DRTAssigningFacilitySequence::gotoNextItem()
 }
 
 
-OFCondition DRTAssigningFacilitySequence::gotoItem(const unsigned long num, OFListIterator(Item *) &iterator)
+OFCondition DRTAssigningFacilitySequence::gotoItem(const size_t num, OFListIterator(Item *) &iterator)
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        unsigned long idx = num + 1;
+        size_t idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -330,12 +330,12 @@ OFCondition DRTAssigningFacilitySequence::gotoItem(const unsigned long num, OFLi
 }
 
 
-OFCondition DRTAssigningFacilitySequence::gotoItem(const unsigned long num, OFListConstIterator(Item *) &iterator) const
+OFCondition DRTAssigningFacilitySequence::gotoItem(const size_t num, OFListConstIterator(Item *) &iterator) const
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        unsigned long idx = num + 1;
+        size_t idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -350,7 +350,7 @@ OFCondition DRTAssigningFacilitySequence::gotoItem(const unsigned long num, OFLi
 }
 
 
-OFCondition DRTAssigningFacilitySequence::gotoItem(const unsigned long num)
+OFCondition DRTAssigningFacilitySequence::gotoItem(const size_t num)
 {
     return gotoItem(num, CurrentItem);
 }
@@ -386,7 +386,7 @@ const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getCurre
 }
 
 
-OFCondition DRTAssigningFacilitySequence::getItem(const unsigned long num, Item *&item)
+OFCondition DRTAssigningFacilitySequence::getItem(const size_t num, Item *&item)
 {
     OFListIterator(Item *) iterator;
     OFCondition result = gotoItem(num, iterator);
@@ -396,7 +396,7 @@ OFCondition DRTAssigningFacilitySequence::getItem(const unsigned long num, Item 
 }
 
 
-DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(const unsigned long num)
+DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(const size_t num)
 {
     OFListIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -406,7 +406,7 @@ DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(const 
 }
 
 
-const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(const unsigned long num) const
+const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(const size_t num) const
 {
     OFListConstIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -416,13 +416,13 @@ const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::getItem(
 }
 
 
-DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::operator[](const unsigned long num)
+DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::operator[](const size_t num)
 {
     return getItem(num);
 }
 
 
-const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::operator[](const unsigned long num) const
+const DRTAssigningFacilitySequence::Item &DRTAssigningFacilitySequence::operator[](const size_t num) const
 {
     return getItem(num);
 }
@@ -445,7 +445,7 @@ OFCondition DRTAssigningFacilitySequence::addItem(Item *&item)
 }
 
 
-OFCondition DRTAssigningFacilitySequence::insertItem(const unsigned long pos, Item *&item)
+OFCondition DRTAssigningFacilitySequence::insertItem(const size_t pos, Item *&item)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
@@ -468,7 +468,7 @@ OFCondition DRTAssigningFacilitySequence::insertItem(const unsigned long pos, It
 }
 
 
-OFCondition DRTAssigningFacilitySequence::removeItem(const unsigned long pos)
+OFCondition DRTAssigningFacilitySequence::removeItem(const size_t pos)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
