+++ dcmrt/libsrc/drtags.cc
@@ -1,13 +1,13 @@
 /*
  *
  *  Copyright (C) 2008-2012, OFFIS e.V. and ICSMED AG, Oldenburg, Germany
- *  Copyright (C) 2013-2016, J. Riesmeier, Oldenburg, Germany
+ *  Copyright (C) 2013-2017, J. Riesmeier, Oldenburg, Germany
  *  All rights reserved.  See COPYRIGHT file for details.
  *
  *  Source file for class DRTApplicatorGeometrySequence
  *
  *  Generated automatically from DICOM PS 3.3-2016e
- *  File created on 2016-11-23 14:23:36
+ *  File created on 2017-01-25 17:55:32
  *
  */
 
@@ -297,7 +297,7 @@ OFBool DRTApplicatorGeometrySequence::isValid() const
 }
 
 
-unsigned long DRTApplicatorGeometrySequence::getNumberOfItems() const
+size_t DRTApplicatorGeometrySequence::getNumberOfItems() const
 {
     return SequenceOfItems.size();
 }
@@ -327,12 +327,12 @@ OFCondition DRTApplicatorGeometrySequence::gotoNextItem()
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::gotoItem(const unsigned long num, OFListIterator(Item *) &iterator)
+OFCondition DRTApplicatorGeometrySequence::gotoItem(const size_t num, OFListIterator(Item *) &iterator)
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        unsigned long idx = num + 1;
+        size_t idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -347,12 +347,12 @@ OFCondition DRTApplicatorGeometrySequence::gotoItem(const unsigned long num, OFL
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::gotoItem(const unsigned long num, OFListConstIterator(Item *) &iterator) const
+OFCondition DRTApplicatorGeometrySequence::gotoItem(const size_t num, OFListConstIterator(Item *) &iterator) const
 {
     OFCondition result = EC_IllegalCall;
     if (!SequenceOfItems.empty())
     {
-        unsigned long idx = num + 1;
+        size_t idx = num + 1;
         iterator = SequenceOfItems.begin();
         const OFListConstIterator(Item *) last = SequenceOfItems.end();
         while ((--idx > 0) && (iterator != last))
@@ -367,7 +367,7 @@ OFCondition DRTApplicatorGeometrySequence::gotoItem(const unsigned long num, OFL
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::gotoItem(const unsigned long num)
+OFCondition DRTApplicatorGeometrySequence::gotoItem(const size_t num)
 {
     return gotoItem(num, CurrentItem);
 }
@@ -403,7 +403,7 @@ const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getCur
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::getItem(const unsigned long num, Item *&item)
+OFCondition DRTApplicatorGeometrySequence::getItem(const size_t num, Item *&item)
 {
     OFListIterator(Item *) iterator;
     OFCondition result = gotoItem(num, iterator);
@@ -413,7 +413,7 @@ OFCondition DRTApplicatorGeometrySequence::getItem(const unsigned long num, Item
 }
 
 
-DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getItem(const unsigned long num)
+DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getItem(const size_t num)
 {
     OFListIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -423,7 +423,7 @@ DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getItem(cons
 }
 
 
-const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getItem(const unsigned long num) const
+const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getItem(const size_t num) const
 {
     OFListConstIterator(Item *) iterator;
     if (gotoItem(num, iterator).good())
@@ -433,13 +433,13 @@ const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::getIte
 }
 
 
-DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::operator[](const unsigned long num)
+DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::operator[](const size_t num)
 {
     return getItem(num);
 }
 
 
-const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::operator[](const unsigned long num) const
+const DRTApplicatorGeometrySequence::Item &DRTApplicatorGeometrySequence::operator[](const size_t num) const
 {
     return getItem(num);
 }
@@ -462,7 +462,7 @@ OFCondition DRTApplicatorGeometrySequence::addItem(Item *&item)
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::insertItem(const unsigned long pos, Item *&item)
+OFCondition DRTApplicatorGeometrySequence::insertItem(const size_t pos, Item *&item)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
@@ -485,7 +485,7 @@ OFCondition DRTApplicatorGeometrySequence::insertItem(const unsigned long pos, I
 }
 
 
-OFCondition DRTApplicatorGeometrySequence::removeItem(const unsigned long pos)
+OFCondition DRTApplicatorGeometrySequence::removeItem(const size_t pos)
 {
     OFCondition result = EC_IllegalCall;
     if (!EmptyDefaultSequence)
